ModularMind/API/services/multimodal/image/processor.py
"""
Görüntü işleme modülü.
"""

import os
import logging
import base64
from typing import Dict, List, Any, Optional, Union, Tuple

logger = logging.getLogger(__name__)

class ImageProcessingError(Exception):
    """Görüntü işleme hatası"""
    pass

class ImageProcessor:
    """
    Görüntü işleme ve özellik çıkarma sınıfı.
    """
    
    def __init__(self):
        """Görüntü işleyici başlatır"""
        self.config = {}
        self.captioning_model = None
        self.feature_model = None
        self.initialized = False
    
    def configure(self, config: Dict[str, Any]) -> None:
        """
        Görüntü işleyiciyi yapılandırır
        
        Args:
            config: Yapılandırma sözlüğü
        """
        self.config = config
    
    def initialize(self) -> bool:
        """
        Gerekli modelleri yükler ve başlatır
        
        Returns:
            bool: Başlatma başarılı mı
        """
        if self.initialized:
            return True
        
        try:
            # Görüntü özellik çıkarma modeli
            feature_model_name = self.config.get("feature_model", "resnet50")
            captioning_model = self.config.get("captioning_model", "blip-large")
            
            # PIL yüklenmişse kontrol et
            try:
                from PIL import Image
                logger.info("PIL başarıyla yüklendi")
            except ImportError:
                logger.error("PIL yüklenmedi. Kurulum: pip install Pillow")
                return False
            
            # Özellik çıkarma modelini yükle
            if feature_model_name == "resnet50":
                try:
                    import torch
                    import torchvision.models as models
                    import torchvision.transforms as transforms
                    
                    # ResNet50 modelini yükle
                    self.feature_model = models.resnet50(pretrained=True)
                    self.feature_model.eval()
                    
                    logger.info("ResNet50 modeli başarıyla yüklendi")
                except ImportError:
                    logger.error("PyTorch veya torchvision yüklenmedi. Kurulum: pip install torch torchvision")
                    return False
            
            # Başlıklandırma modelini yükle
            if captioning_model == "blip-large":
                try:
                    # BLIP modelini buraya ekleyin
                    # Not: Bu örnek kod içi bu kütüphane kurulumu gerektiriyor
                    # from lavis.models import load_model_and_preprocess
                    # self.captioning_model, self.vis_processors, _ = load_model_and_preprocess(
                    #     name="blip_caption", model_type="large_coco", is_eval=True, device="cpu"
                    # )
                    
                    # Örnek için OpenAI API kullanalım
                    import openai
                    self.captioning_model = "openai"
                    
                    logger.info("Başlıklandırma modeli başarıyla yapılandırıldı")
                except ImportError:
                    logger.warning("BLIP modeli yüklenemedi, OpenAI API kullanılacak")
                    self.captioning_model = "openai"
            
            self.initialized = True
            return True
        except Exception as e:
            logger.error(f"Görüntü işleme modelleri başlatılamadı: {str(e)}")
            return False
    
    def process(self, image_path: str, options: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
        """
        Görüntüyü işler ve analiz eder
        
        Args:
            image_path: Görüntü dosyası yolu
            options: İşleme seçenekleri
            
        Returns:
            Dict[str, Any]: İşleme sonuçları
        """
        if not self.initialized:
            if not self.initialize():
                return {"error": "Görüntü işleme başlatılamadı"}
        
        options = options or {}
        extract_features = options.get("extract_features", True)
        generate_caption = options.get("generate_caption", True)
        
        result = {
            "image_path": image_path,
            "format": None,
            "width": None,
            "height": None
        }
        
        try:
            # Görüntü bilgilerini al
            from PIL import Image
            img = Image.open(image_path)
            
            result["format"] = img.format
            result["width"], result["height"] = img.size
            
            # Özellik çıkarma
            if extract_features:
                features = self.extract_features(image_path)
                result.update(features)
            
            # Başlıklandırma
            if generate_caption:
                caption = self.caption_image(image_path)
                if caption:
                    result["caption"] = caption
            
            return result
        except Exception as e:
            logger.error(f"Görüntü işleme hatası: {str(e)}")
            return {"error": str(e)}
    
    def extract_features(self, image_path: str, options: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
        """
        Görüntüden özellikler çıkarır
        
        Args:
            image_path: Görüntü dosyası yolu
            options: Çıkarma seçenekleri
            
        Returns:
            Dict[str, Any]: Çıkarılan özellikler
        """
        if not self.initialized:
            if not self.initialize():
                return {"error": "Görüntü işleme başlatılamadı"}
        
        options = options or {}
        
        try:
            # ResNet50 ile özellik çıkarma
            import torch
            import torchvision.transforms as transforms
            from PIL import Image
            
            # Görüntüyü yükle ve dönüştür
            transform = transforms.Compose([
                transforms.Resize(256),
                transforms.CenterCrop(224),
                transforms.ToTensor(),
                transforms.Normalize(
                    mean=[0.485, 0.456, 0.406],
                    std=[0.229, 0.224, 0.225]
                )
            ])
            
            img = Image.open(image_path).convert("RGB")
            img_tensor = transform(img).unsqueeze(0)
            
            # Özellikleri çıkar
            with torch.no_grad():
                features = self.feature_model(img_tensor)
                features = features.squeeze().tolist()
            
            return {
                "features": features,
                "feature_dim": len(features),
                "model": "resnet50"
            }
        except Exception as e:
            logger.error(f"Görüntü özellik çıkarma hatası: {str(e)}")
            return {"error": str(e)}
    
    def caption_image(self, image_path: str, options: Optional[Dict[str, Any]] = None) -> Optional[str]:
        """
        Görüntüyü başlıklandırır
        
        Args:
            image_path: Görüntü dosyası yolu
            options: Başlıklandırma seçenekleri
            
        Returns:
            Optional[str]: Oluşturulan başlık
        """
        if not self.initialized:
            if not self.initialize():
                return None
        
        options = options or {}
        
        try:
            if self.captioning_model == "openai":
                # OpenAI API ile başlıklandırma
                import openai
                import os
                
                api_key = os.environ.get("OPENAI_API_KEY")
                if not api_key:
                    logger.error("OPENAI_API_KEY çevresel değişkeni ayarlanmamış")
                    return None
                
                # Görüntüyü base64'e çevir
                with open(image_path, "rb") as image_file:
                    base64_image = base64.b64encode(image_file.read()).decode('utf-8')
                
                # API isteği yap
                client = openai.OpenAI(api_key=api_key)
                response = client.chat.completions.create(
                    model="gpt-4-vision-preview",
                    messages=[
                        {
                            "role": "user",
                            "content": [
                                {"type": "text", "text": "Bu görseli kısaca açıkla."},
                                {
                                    "type": "image_url",
                                    "image_url": {"url": f"data:image/jpeg;base64,{base64_image}"}
                                }
                            ]
                        }
                    ],
                    max_tokens=300
                )
                
                caption = response.choices[0].message.content
                return caption
            else:
                # BLIP veya başka bir model kullanarak başlıklandırma
                return "Bu özellik henüz uygulanmadı"
        except Exception as e:
            logger.error(f"Görüntü başlıklandırma hatası: {str(e)}")
            return None