requirements.txt
# Temel gereksinimler
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0
typer>=0.9.0

# Veritabanı ve vektör indeksleme
hnswlib>=0.7.0
faiss-cpu>=1.7.4
qdrant-client>=1.4.0
weaviate-client>=3.24.1
pinecone-client>=2.2.2
redis>=4.6.0
pymongo>=4.5.0
elasticsearch>=8.9.0

# Embedding modelleri
openai>=1.3.0
cohere>=4.32
sentence-transformers>=2.2.2
transformers>=4.34.0
huggingface-hub>=0.17.3

# Belge işleme
langdetect>=1.0.9
pdfplumber>=0.10.2
python-docx>=0.8.11
pytesseract>=0.3.10
pillow>=10.0.0
bs4>=0.0.1
requests>=2.31.0
lxml>=4.9.3

# Çoklu modalite
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2
ffmpeg-python>=0.2.0
soundfile>=0.12.1
transformers>=4.34.0

# İzleme ve performans
prometheus-client>=0.17.1
prometheus-fastapi-instrumentator>=6.1.0
structlog>=23.1.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0

# Test ve geliştirme araçları
pytest>=7.4.2
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.9.1
isort>=5.12.0
mypy>=1.5.1